requirements.txt
customtkinter
Pillow
pyperclip
inflect
